로그인 화면 배경화면 변경
사용자 관리화면 추가
출퇴근 기록 화면 추가
CLAUDE.md 파일 최적화

+++ src/main/java/io/github/mskim/comm/cms/dto/UserDTO.java
@@ -9,7 +9,7 @@ public class UserDTO extends BaseDTO{
     private String loginId;         // 로그인 아이디
     private String password;        // 비밀번호
     private String name;            // 이름
-    private int annualLeaveDays;    // 연차
+    private Double annualLeaveDays;    // 연차
     private String role;            // 권한
     private String position;            // 직급
 }
